moved basic_auth.py to core
