+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/app.py
@@ -12,7 +12,7 @@ from stac_fastapi.core.core import (
 )
 from stac_fastapi.core.extensions import QueryExtension
 from stac_fastapi.core.session import Session
-from stac_fastapi.elasticsearch.basic_auth import apply_basic_auth
+from stac_fastapi.core.basic_auth import apply_basic_auth
 from stac_fastapi.elasticsearch.config import ElasticsearchSettings
 from stac_fastapi.elasticsearch.database_logic import (
     DatabaseLogic,
