+++ stac_fastapi/opensearch/stac_fastapi/opensearch/app.py
@@ -21,7 +21,7 @@ from stac_fastapi.extensions.core import (
     TransactionExtension,
 )
 from stac_fastapi.extensions.third_party import BulkTransactionExtension
-from stac_fastapi.opensearch.basic_auth import apply_basic_auth
+from stac_fastapi.core.basic_auth import apply_basic_auth
 from stac_fastapi.opensearch.config import OpensearchSettings
 from stac_fastapi.opensearch.database_logic import (
     DatabaseLogic,
