+++ stac_fastapi/tests/conftest.py
@@ -16,9 +16,9 @@ from stac_fastapi.core.core import (
     TransactionsClient,
 )
 from stac_fastapi.core.extensions import QueryExtension
+from stac_fastapi.core.basic_auth import apply_basic_auth
 
 if os.getenv("BACKEND", "elasticsearch").lower() == "opensearch":
-    from stac_fastapi.opensearch.basic_auth import apply_basic_auth
     from stac_fastapi.opensearch.config import AsyncOpensearchSettings as AsyncSettings
     from stac_fastapi.opensearch.config import OpensearchSettings as SearchSettings
     from stac_fastapi.opensearch.database_logic import (
@@ -36,7 +36,6 @@ else:
         create_collection_index,
         create_index_templates,
     )
-    from stac_fastapi.elasticsearch.basic_auth import apply_basic_auth
 
 from stac_fastapi.extensions.core import (  # FieldsExtension,
     ContextExtension,
